Moved the functions to the proc macro, still missing imports
