+++ crates/hyperprocess_macro/Cargo.toml
@@ -10,3 +10,17 @@ proc-macro = true
 syn = { version = "2.0", features = ["full"] }
 quote = "1.0"
 proc-macro2 = "1.0"
+
+anyhow = "1.0"
+hyperware_process_lib = { version = "1.0.2", features = ["logging"] }
+process_macros = "0.1.0"
+futures-util = "0.3"
+serde = { version = "1.0", features = ["derive"] }
+rmp-serde = "1.0"
+serde_json = "1.0"
+serde_derive = "1.0"
+wit-bindgen = "0.36.0"
+once_cell = "1.20.2"
+paste = "1.0"
+uuid = { version = "1.0", features = ["v4"] }
+hyperware_app_common = { path = "../hyperware_app_common" }
