+++ hyperdriver/Cargo.toml
@@ -8,6 +8,7 @@ publish = false
 anyhow = "1.0"
 hyperware_process_lib = { version = "1.0.2", features = ["logging"] }
 process_macros = "0.1.0"
+futures-util = "0.3"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 wit-bindgen = "0.36.0"
